feat(structure): initial implementation of OpenXR camera rig
The OpenXR camera rig provides a prefab that should work with any
supported OpenXR vendor plugin. It needs unity to be configured for
each vendor plugin and there are samples with the package to provide
access to OpenXR extensions. At the moment the only extension supported
is Passthrough on a basic level.

The package also contains an editor helper to allow quick switching
between OpenXR profile settings in Unity as it would seem Unity does
not offer this functionality, and even in Unity 6 with Build Profiles
it does not seem supported.