fix(PassthroughProcessor): ensure source manager logic is called
The SourceManager in the passthrough manager is supposed to allow an
extended passthrough processor to halt the AR Foundation processing
done by the Passthrough Manager. However, this logic was not being
set so it was never being limited.

This has now been fixed so the passthrough processors do actually
now control the state of the SourceManager.